Fix crash on iOS 7 and 8 with iOS 9+ API calls
This fixes a crash on iOS 7 and 8 mapview properties
showsCompass, showsScale or showsTraffic are set.
Thse properties are only available on iOS 9+.

Similar to the origin react-native MapView (RCTMap.m)
this checks if the super class (MKMapView) has selectors
to set and get these parameters. If not it does nothing or
return false.